Adapt master datastore
- adapt master datastore to do not invoke datastore
operation and listener action when RPC is called and
no data are changed compared to original datastore state

+++ sxp-controller/src/main/java/org/opendaylight/sxp/controller/core/SxpRpcServiceImpl.java
@@ -541,7 +541,8 @@ public class SxpRpcServiceImpl implements SxpControllerService, AutoCloseable {
             }
 
             final MasterDatabaseInf masterDatabase = getMasterDatabase(nodeId, input.getDomainName());
-            if (input.getBinding() != null && !input.getBinding().isEmpty() && masterDatabase != null) {
+            if (input.getBinding() != null && !input.getBinding().isEmpty()
+                    && masterDatabase != null && masterDatabase.getBindings() != null) {
                 final List<MasterDatabaseBinding> bindingsToBeRemoved = transformBindings(input.getBinding());
                 if (!bindingsToBeRemoved.isEmpty()) {
                     final List<MasterDatabaseBinding> deleted = masterDatabase.deleteBindingsLocal(bindingsToBeRemoved);
